added full implementation of dao

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `books` (
 `id` BIGINT NOT NULL AUTO_INCREMENT,
  `title` VARCHAR(250),
- `price` INT,
+ `price` DECIMAL,
  PRIMARY KEY (`id`));
